refactor(core): move UserProfile, LLM prompts and response parsing into core
Move UserProfile from open-course-config to open_course_core::profile so it
can be shared with external consumers; config re-exports it, keeping the
config.json format and the open_course_config::profile::UserProfile path
unchanged (serde attributes moved verbatim).

Move crates/llm/src/prompts.rs and crates/llm/src/parse.rs to
open_course_core::llm::{prompts, parse} so a server crate can reuse the
prompt builders and response parsing via a git dependency on core alone,
without the LLM client stack (rig-core, reqwest, tokio). Items become pub;
LlmResponse moves to open_course_core::llm::response.

open-course-llm re-exports both modules, so all existing call sites in
service/cli and the open_course_llm::{prompts, parse} paths are unchanged.
CLI behavior is unchanged.
