+++ crates/llm/src/lib.rs
@@ -5,15 +5,20 @@ pub mod debug_log;
 pub mod diagnostics;
 pub mod factory;
 pub mod model_listing;
-pub mod parse;
 pub mod pipeline;
-pub mod prompts;
 pub mod provider;
 pub mod result;
 pub mod retry;
 pub mod streaming;
 pub mod topic_review;
 pub mod transport;
+
+// Prompt builders and response parsing live in `open-course-core` so external
+// consumers (e.g. the server) can reuse them without the LLM client stack.
+// Re-exported here to keep the existing `open_course_llm::{prompts, parse}`
+// paths working.
+pub use open_course_core::llm::parse;
+pub use open_course_core::llm::prompts;
 
 pub use result::LlmResult;
